backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
anthropic==0.42.0
openai>=1.60.0
google-genai>=1.0.0
# sentence-transformers>=3.0.0  # RAG - pulls in PyTorch (~3-4GB)
# faiss-cpu==1.9.0              # RAG
python-dotenv==1.0.1
websockets==14.1
python-multipart==0.0.19
aiofiles==24.1.0
py-mini-racer==0.6.0
httpx==0.28.1
mcp>=1.0.0
numpy==2.2.1
watchdog>=4.0.0
feedparser>=6.0.0
croniter>=6.0.0
python-lsp-server[rope]
mammoth>=1.8.0
pyodbc>=5.0.0
psycopg2-binary>=2.9.0
azure-cosmos>=4.7.0
azure-storage-blob>=12.19.0
Pillow>=10.0.0
PyJWT>=2.8.0
bcrypt>=4.1.0
